requirements.txt
django~=4.0.0
djangorestframework~=3.13.0
django-cors-headers~=3.10.0
wheel==0.37.1
dj-rest-auth==2.2.4
requests==2.27.1
django-allauth~=0.48.0
drf-spectacular~=0.21.0
psycopg2-binary==2.9.3
whitenoise==6.0.0
boto3==1.21.23
botocore==1.24.23
django-ses==3.0.1
sqlalchemy==1.4.40
boto3==1.21.23
python-decouple==3.6
pytz==2021.3
tzdata==2022.7
celery==5.2.7
redis==4.4.2
djangorestframework-simplejwt==5.2.2